backend/requirements.txt
pyswisseph
numpy
fastapi
uvicorn[standard]
pydantic
pytz
